Menu domain layer was added

+++ app/src/main/java/com/example/coffeeapp/domain/main/shops/ShopLocation.kt
@@ -3,7 +3,7 @@ package com.example.coffeeapp.domain.main.shops
 import com.example.coffeeapp.data.main.shops.remote.Point
 
 data class ShopLocation(
-    val id: Double,
+    val id: Int,
     val name: String,
     val point: Point,
     val distance: Double? = null
